Moved Price endpoint to RidesController. Deleted PriceController
