+++ SmartCab.Backend/Api/Controllers/RidesController.cs
@@ -67,5 +67,39 @@ namespace Api.Controllers
             var response = await _rideService.AddRideAsync(request, customerId);
             return Ok(response);
         }
+
+        /// <summary>
+        /// Calculates and returns the price for a taxi ride based on two addresses.
+        /// </summary>
+        /// <remarks>
+        /// Currently now authorization is required to make this request.
+        /// </remarks>
+        /// <param name="authorization">A valid JWT token.</param>
+        /// <param name="request">Consists of an start address and an end address.</param>
+        /// <returns> the price</returns>
+        /// <response code="400">If the supplied request wasn't valid.</response>
+        /// <response code="401">If the token was expired</response>
+        /// <response code="500">If an internal server error occured.</response>
+        [Authorize(Roles = nameof(Customer))]
+        [Produces("application/json")]
+        [Route("[action]")]
+        [ProducesResponseType(typeof(decimal), StatusCodes.Status200OK)]
+        [HttpPost]
+        public async Task<IActionResult> Price([FromHeader] string authorization, [FromBody] PriceRequest request)
+        {
+            //Get the customerId, stored as a claim in the token
+            var customerId = User.Claims.FirstOrDefault(x => x.Type == Constants.UserIdClaim)?.Value;
+
+            //should never happen
+            if (string.IsNullOrEmpty(customerId))
+            {
+                throw new UserIdInvalidException(
+                    $"The supplied JSON Web Token does not contain a valid value in the '{ Constants.UserIdClaim }' claim.");
+            }
+
+            var calculatedPrice = await
+                _rideService.CalculatePriceAsync(request.StartAddress, request.EndAddress, request.RideType);
+            return Ok(calculatedPrice);
+        }
     }
 }
